yet more cleanup and moving things around for tidyiness and testability
