+++ backend/mr-db-sql/pom.xml
@@ -76,18 +76,17 @@
     </dependency>
 
     <dependency>
+      <groupId>io.featurehub.messaging</groupId>
+      <artifactId>mr-messaging-channel</artifactId>
+      <version>1.1-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
       <groupId>io.featurehub.common</groupId>
       <artifactId>common-db-test</artifactId>
       <version>1.1-SNAPSHOT</version>
       <scope>test</scope>
     </dependency>
-
-    <dependency>
-      <groupId>io.featurehub.messaging</groupId>
-      <artifactId>mr-messaging-processor</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-
   </dependencies>
 
   <build>
