+++ backend/mr-eventing/pom.xml
@@ -81,8 +81,6 @@
       <artifactId>eventing-kinesis</artifactId>
       <version>1.1-SNAPSHOT</version>
     </dependency>
-
-
     <dependency>
       <groupId>io.featurehub.common</groupId>
       <artifactId>common-db-test</artifactId>
@@ -90,13 +88,8 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>io.featurehub.api</groupId>
-      <artifactId>messaging-api</artifactId>
-      <version>1.1-SNAPSHOT</version>
-    </dependency>
-    <dependency>
       <groupId>io.featurehub.messaging</groupId>
-      <artifactId>mr-messaging-processor</artifactId>
+      <artifactId>mr-messaging-channel</artifactId>
       <version>1.1-SNAPSHOT</version>
     </dependency>
   </dependencies>
