+++ backend/mr-eventing/src/main/kotlin/io/featurehub/mr/events/EventingFeature.kt
@@ -7,7 +7,7 @@ import io.featurehub.db.publish.nats.NatsDachaEventingFeature
 import io.featurehub.events.CloudEventsFeature
 import io.featurehub.events.kinesis.KinesisEventFeature
 import io.featurehub.events.pubsub.GoogleEventFeature
-import io.featurehub.messaging.ManagementRepositoryMessagingFeature
+import io.featurehub.messaging.MessagingPublishingFeature
 import io.featurehub.mr.events.common.CacheBroadcast
 import io.featurehub.mr.events.common.CacheSource
 import io.featurehub.mr.events.common.CloudEventCacheBroadcaster
@@ -24,27 +24,25 @@ import org.glassfish.jersey.internal.inject.AbstractBinder
 
 class EventingFeature : Feature {
   override fun configure(context: FeatureContext): Boolean {
-    var amPublishing = GoogleEventFeature.isEnabled() || NatsDachaEventingFeature.isEnabled() || KinesisEventFeature.isEnabled()
+    var amPublishing =
+      GoogleEventFeature.isEnabled() || NatsDachaEventingFeature.isEnabled() || KinesisEventFeature.isEnabled()
 
     context.register(CloudEventsFeature::class.java)
     context.register(PubsubMRFeature::class.java) // this will in fact not register anything if it is not enabled
-    context.register(KinesisMRFeature::class.java)
+    context.register(KinesisMRFeature::class.java) // this will in fact not register anything if it is not enabled
+    context.register(NatsDachaEventingFeature::class.java) // this will in fact not register anything if it is not enabled
 
-    if (NatsDachaEventingFeature.isEnabled()) {
-      context.register(NatsDachaEventingFeature::class.java)
-    }
+    context.register(MessagingPublishingFeature::class.java)
 
-    context.register(ManagementRepositoryMessagingFeature::class.java)
-
-
-    context.register(object: AbstractBinder() {
+    context.register(object : AbstractBinder() {
       override fun configure() {
         bind(FeatureUpdateListenerImpl::class.java).to(FeatureUpdateListener::class.java).`in`(Immediate::class.java)
 
         if (amPublishing) {
           // the broadcaster will determine if dacha2 is enabled and not publish to that channel if not
           bind(CloudEventCacheBroadcaster::class.java).to(CacheBroadcast::class.java).`in`(Singleton::class.java)
-          bind(DbCacheSource::class.java).to(CacheSource::class.java).to(CacheApi::class.java).to(CacheRefresherApi::class.java)
+          bind(DbCacheSource::class.java).to(CacheSource::class.java).to(CacheApi::class.java)
+            .to(CacheRefresherApi::class.java)
             .`in`(
               Singleton::class.java
             )
