+++ backend/mr-messaging-channel/src/main/kotlin/io/featurehub/messaging/MessagingPublishingFeature.kt
@@ -1,5 +1,6 @@
 package io.featurehub.messaging
 
+import io.featurehub.db.messaging.FeatureMessagingPublisher
 import io.featurehub.lifecycle.LifecycleListeners
 import io.featurehub.messaging.service.FeatureMessagingCloudEventInitializer
 import io.featurehub.messaging.service.FeatureMessagingCloudEventPublisher
@@ -15,7 +16,11 @@ class MessagingPublishingFeature : Feature {
     ctx.register(object : AbstractBinder() {
 
       override fun configure() {
-        bind(FeatureMessagingCloudEventPublisherImpl::class.java).to(FeatureMessagingCloudEventPublisher::class.java)
+        // we have split this into two so db-sql doesn't have to depend on this
+        // artifact to build, it separates cloud eventing from publishing
+        bind(FeatureMessagingCloudEventPublisherImpl::class.java)
+          .to(FeatureMessagingPublisher::class.java)
+          .to(FeatureMessagingCloudEventPublisher::class.java)
           .`in`(Singleton::class.java)
       }
