+++ backend/mr-messaging-channel/src/main/kotlin/io/featurehub/messaging/service/FeatureMessagingCloudEventPublisher.kt
@@ -1,0 +1,13 @@
+package io.featurehub.messaging.service
+
+import io.featurehub.events.DynamicCloudEventDestination
+
+
+interface FeatureMessagingCloudEventPublisher {
+  /**
+   * This is set up by the initializer
+   */
+  fun setHooks(hooks: List<DynamicCloudEventDestination>)
+
+  val isEnabled: Boolean
+}
