+++ pom.xml
@@ -90,7 +90,7 @@
     <module>backend/mr-eventing</module>
     <module>backend/mr-eventing-common</module>
     <module>backend/messaging-api</module>
-    <module>backend/messaging-channel</module>
+    <module>backend/mr-messaging-channel</module>
     <module>backend/mr-eventing-nats</module>
     <module>backend/edge-feature-update-async-api</module>
     <module>backend/eventing-nats</module>
